Allow different rules for climbing (and jumping) with backpacks (by JMich) - ja2_options.ini: MAX_BACKPACK_WEIGHT_TO_CLIMB, USE_GLOBAL_BACKPACK_SETTINGS
MAX_BACKPACK_WEIGHT_TO_CLIMB is the maximum weight a backpack can weigh and still allow you to climb with. 
The sBackpackWeightModifier in items.xml can modify that number by either making it harder (positive number modifier) 
or easier (negative number modifier). If the sBackpackWeightToClimb is -1 (default) then the option is unused.
 
If USE_GLOBAL_BACKPACK_SETTINGS is TRUE (default), it uses the same rules for all backpacks, while if FALSE checks for 
the fAllowClimbing tag in Items.xml, and if that is true you can climb with the backpack.
 
Both options can be used at the same time, so you may climb with (for example) a malice backpack that weighs less than 5kg
(or 50 hg) but not climb with a radio set.

git-svn-id: https://ja2svn.mooo.com/source/ja2_v1.13_data@2008 4f8fa57e-7814-0410-bad4-adc449f26b7c